Move remaining ActiveRecord models to `app/models`
This makes the app look more like a normal Rails app.
